requirement.txt
alpaca-trade-api
websocket-client
stable-baselines3
gym
numpy
pandas
TA-Lib
textblob
requests
joblib
torch
scikit-learn
matplotlib
seaborn
